开始行驶，结束行驶调通接口

+++ app/src/main/java/me/xucan/safedrive/bean/DriveRecord.java
@@ -31,7 +31,7 @@ public class DriveRecord implements Serializable{
     private String endPlace;
 
     @Column(name = "distance")
-    private int distance;
+    private float distance;
 
     @Column(name = "safetyIndex")
     private int safetyIndex;
@@ -92,11 +92,11 @@ public class DriveRecord implements Serializable{
         this.endPlace = endPlace;
     }
 
-    public int getDistance() {
+    public float getDistance() {
         return distance;
     }
 
-    public void setDistance(int distance) {
+    public void setDistance(float distance) {
         this.distance = distance;
     }
 }
